refactor(examples): move workflow demo under experimental
