+++ examples/experimental/workflow/hello_world/stages.py
@@ -6,7 +6,7 @@
 from collections.abc import Mapping
 from typing import Any
 
-from dynamo.workflow import StageContext, StageContract
+from dynamo.experimental.workflow import StageContext, StageContract
 
 
 class HelloStage:
